Now using the shorter of the two radii a segment is a part of rather than the average.
I was finding that for many roads, using the average resulted in just the apex
of curves being identifed under any reasonable threashold for maximum radius.
Using the shorter of the two radii may make the curvature for the series of segments
a little more "jittery" than the average, but does well at highlighting the full
length of medium-radius turns.