Fixed collector error where ways joined in one route could be reversed while joining them to other ways.
Vermont's 'VT 17' route includes several segements that are also part of the
'VT 116' route. The 'VT 17 ways would get reversed and added correctly, but
because the same way object in memory was also being joined to 'VT 116'
it would get reversed again to join it to 'VT 116', leaving in in an invalid
state in the 'VT 17' sequence where the final refs of on way wouldn't match the
starting refs of the next way.

#21