Revert addition of a `lib/` folder.
After researching this some more:
 https://stackoverflow.com/questions/193161/what-is-the-best-project-structure-for-a-python-application
 https://intermediate-and-advanced-software-carpentry.readthedocs.io/en/latest/structuring-python.html
it turns out that the norm for Python programs is to have the main module
directory at the top level and not nest it under `lib/`, `src/` or something
else. This makes setting up the environment a bit and should open the door
for packaging curvature for installation via easy_install or pip.