Convert from imposm.parser to osmium for loading roads out of the OSM PBF.
Osmium returns the geometry with the ways so we don't have to do the multi-pass
loading that was required with imposm.parser.

The working-data structure is remaining the same. We can't read multiple
times from the osmium objects anyway, but even still, this keeps the rest
of the tool chain from having to change.