Squash segment curvature when transitioning from oneway to two-way roads.
This fixes the issue in #55 where "chained" streets that are modeled 
as a single way in undivided segments and pairs of oneway ways in 
divided segments. Most transitions I've looked at have been about 
15-20m long, so a threshold of 30m should be enough to avoid these
phantom curves while not being so agressive as to squash the curvature
of an entire motorway entrance ramp.

By also filtering on the transition from junction=roundabout to
non-roundabout ways, we can avoid including a portion of roundabout 
entrance/exit ways. These spots coincide with other conflict zones
and shouldn't be included in the "fun curves" amount, even though
their geometry *is* curvy.